+++ sdk/androidsdk/src/main/java/com/loginradius/androidsdk/api/SocialInterfaceAPI.java
@@ -12,16 +12,15 @@ import io.reactivex.android.schedulers.AndroidSchedulers;
 import io.reactivex.observers.DisposableObserver;
 import io.reactivex.schedulers.Schedulers;
 import retrofit2.HttpException;
-import retrofit2.http.Header;
 
 
-public class SocailInterfaceAPI {
+public class SocialInterfaceAPI {
 
 
 	public void getResponse(LoginParams value, final AsyncHandler<SocialInterface> handler) {
 
 		ApiInterface apiService = RestRequest.getClientCdn().create(ApiInterface.class);
-		apiService.getSocailProviderInterface(Endpoint.API_V2_SOCIALINTERFACE_URL+value.apikey+".json").subscribeOn(Schedulers.io()).observeOn(AndroidSchedulers.mainThread())
+		apiService.getSocialProviderInterface(Endpoint.API_V2_SOCIALINTERFACE_URL+value.apikey+".json").subscribeOn(Schedulers.io()).observeOn(AndroidSchedulers.mainThread())
 				.subscribe(new DisposableObserver<SocialInterface>() {
 					@Override
 					public void onComplete() {